Don't spin up new context on every command.
As designed original, whenever a client submitted a job,
ZeroMQPubServerChannel was spinning up a new context, and zmq PUSH
socket to submit to the publish worker. While we where calling close()
and term() on the socket and context, we where not setting linger. The
effect is under heavy load (lots of outbound jobs and/or lots of
minions) the publisher has to pull the job from the socket BEFORE the
python garbage collector decides to gobble up the PUSH socket. This
results in jobs returning to the client with JID’s and the job never
making it to the minion.

The fix in this patch is to simply spin up the PUSH socket at init
time, keep it around, and tear down when we are being destroyed. This
also increases performance because the creation time for a zmq push
socket is not small.